+++ src/components/form/fields/phoneNumberField.tsx
@@ -182,7 +182,7 @@ const phoneNumberField = (
                 }
             },
             unbind: formValue => {
-                return (typeof formValue === 'object' && 'raw' in formValue ? formValue.raw : formValue) ?? null
+                return (typeof formValue === 'object' && 'raw' in formValue ? formValue.raw : formValue)
             }
         },
         validator: new Validator<Value>({